:truck: Move enums to data_safe_haven.types
