+++ data_safe_haven/infrastructure/programs/sre/dns_server.py
@@ -8,15 +8,17 @@ from pulumi_azure_native import containerinstance, network, resources
 
 from data_safe_haven.functions import b64encode
 from data_safe_haven.infrastructure.common import (
-    AzureDnsZoneNames,
-    NetworkingPriorities,
-    PermittedDomains,
-    Ports,
     SREDnsIpRanges,
     SREIpRanges,
     get_ip_address_from_container_group,
 )
 from data_safe_haven.resources import resources_path
+from data_safe_haven.types import (
+    AzureDnsZoneNames,
+    NetworkingPriorities,
+    PermittedDomains,
+    Ports,
+)
 from data_safe_haven.utility import FileReader
