+++ data_safe_haven/types/__init__.py
@@ -9,19 +9,32 @@ from .annotated_types import (
     TimeZone,
     UniqueList,
 )
-from .enums import DatabaseSystem, SoftwarePackageCategory
+from .enums import (
+    AzureDnsZoneNames,
+    DatabaseSystem,
+    FirewallPriorities,
+    NetworkingPriorities,
+    PermittedDomains,
+    Ports,
+    SoftwarePackageCategory,
+)
 from .types import PathType
 
 __all__ = [
+    "AzureDnsZoneNames",
     "AzureLocation",
     "AzureLongName",
     "AzureVmSku",
     "DatabaseSystem",
     "EmailAddress",
+    "FirewallPriorities",
     "Fqdn",
     "Guid",
     "IpAddress",
+    "NetworkingPriorities",
     "PathType",
+    "PermittedDomains",
+    "Ports",
     "SoftwarePackageCategory",
     "TimeZone",
     "UniqueList",
